Update to new package conventions
