--- packages/package_info/example/android/app/src/main/java/dev/flutter/plugins/packageinfoexample/MainActivity.java
@@ -1,16 +1,0 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-package dev.flutter.plugins.packageinfoexample;
-
-import dev.flutter.plugins.packageinfo.PackageInfoPlugin;
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.engine.FlutterEngine;
-
-public class MainActivity extends FlutterActivity {
-  @Override
-  public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    flutterEngine.getPlugins().add(new PackageInfoPlugin());
-  }
-}
