+++ packages/package_info/pubspec.yaml
@@ -7,7 +7,7 @@ version: 0.4.0+8
 
 flutter:
   plugin:
-    androidPackage: dev.flutter.plugins.packageinfo
+    androidPackage: io.flutter.plugins.packageinfo
     iosPrefix: FLT
     pluginClass: PackageInfoPlugin
